+++ inst/shinydashboard.js
@@ -43,16 +43,21 @@ var ensureActivatedTab = function() {
   // If there's a `data-start-selected` attribute and we can find a tab with
   // that name, activate it.
   var $startTab = $tablinks.filter("[data-start-selected='1']");
-  if ($startTab.length !== 0) {
-    $startTab.tab("show");
-    return;
+  if ($startTab.length === 0) {
+    // If no tab starts selected, use the first one, if present
+    $startTab = $tablinks.first();
   }
 
-  // If we got this far, just activate the first tab.
-  if (! $tablinks.parent("li").hasClass("active") ) {
-    $tablinks.first().tab("show");
+  // If there are no tabs, $startTab.length will be 0.
+  if ($startTab.length !== 0) {
+    $startTab.tab("show");
+
+    // This is indirectly setting the value of the Shiny input by setting
+    // an attribute on the html element it is bound to. We cannot use the
+    // inputBinding's setValue() method here because this is called too
+    // early (before Shiny has fully initialized)
     $(".sidebarMenuSelectedTabItem").attr("data-value",
-      $tablinks.first().attr("data-value"));
+      $startTab.attr("data-value"));
   }
 };
 
@@ -68,17 +73,50 @@ if ($("section.sidebar").data("disable")) {
   $(".navbar > .sidebar-toggle").hide();
 }
 
-// Whenever the sidebar expand/collapse button is clicked:
-$(document).on("click", ".sidebar-toggle", function() {
+// Get the correct value for `input$sidebarCollapsed`, depending on
+// whether or not the left offset on the sidebar is negative (hidden
+// - so `input$sidebarCollapsed` should be TRUE) or 0 (shown - so
+// `input$sidebarCollapsed` should be FALSE). That we know of,
+// `$(".main-sidebar").is(":visible")` is always true, so there is
+// no need to check for that.
+var sidebarCollapsedValue = function() {
+  if ($(".main-sidebar").offset().left < 0) return(true);
+  else return(false);
+};
+
+// Whenever the sidebar changes from collapsed to expanded and vice versa,
+// call this function, so that we can trigger the resize event on the rest
+// of the window and also update the value for the sidebar's input binding.
+var sidebarChange = function() {
   // 1) Trigger the resize event (so images are responsive and resize)
   $(window).trigger("resize");
 
   // 2) Update the value for the sidebar's input binding
   var $obj = $('.main-sidebar.shiny-bound-input');
   var inputBinding = $obj.data('shiny-input-binding');
-  inputBinding.toggleValue($obj);
+  inputBinding.setValue($obj, sidebarCollapsedValue());
   $obj.trigger('change');
-});
+};
+
+// Whenever the sidebar finishes a transition (which it does every time it
+// changes from collapsed to expanded and vice versa), call sidebarChange()
+$(".main-sidebar").on(
+  'webkitTransitionEnd otransitionend oTransitionEnd msTransitionEnd transitionend',
+    sidebarChange);
+
+// This fixes an edge case: when the app starts up with the sidebar already
+// collapsed (either because the screen is small or because
+// `dashboardSidebar(disable = TRUE`), make sure that `input$sidebarCollapsed`
+// is set to `FALSE`. Whenever this is the case, `$(".main-sidebar").offset().left`
+// is negative. That we know of, `$(".main-sidebar").is(":visible")` is always
+// true, so there is no need to check for that.
+if ($(".main-sidebar").offset().left < 0) {
+  // This is indirectly setting the value of the Shiny input by setting
+  // an attribute on the html element it is bound to. We cannot use the
+  // inputBinding's setValue() method here because this is called too
+  // early (before Shiny has fully initialized)
+  $(".main-sidebar").attr("data-collapsed", "true");
+}
 
 // Whenever we expand a menuItem (to be expandable, it must have children),
 // update the value for the expandedItem's input binding (this is the